# src/main/java/com/cmcu/itstudy/dto/auth/ForgotPasswordRequestDto.java
package com.cmcu.itstudy.dto.auth;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ForgotPasswordRequestDto {

    private String email;
}