# src/main/java/com/cmcu/itstudy/dto/common/ErrorResponseDto.java
package com.cmcu.itstudy.dto.common;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ErrorResponseDto {

    private String code;
    private String message;
}