# src/main/java/com/cmcu/itstudy/dto/role/AssignPermissionRequestDto.java
package com.cmcu.itstudy.dto.role;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class AssignPermissionRequestDto {

    private String permissionId;
}