# src/main/java/com/cmcu/itstudy/dto/user/AssignRoleRequestDto.java
package com.cmcu.itstudy.dto.user;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class AssignRoleRequestDto {

    private String roleId;
}